Move WriteOnlyPropertyInspectionTests for QuickFixes into relevant files under RD.QuickFixes
